requirements.txt
python-dotenv
transformers
torch
huggingface-hub